UPDATE: posts business logic update

+++ src/posts/services.py
@@ -1,34 +1,28 @@
-import json
 from typing import List, Optional
-from fastapi import Depends, HTTPException, logger
-from sqlalchemy import insert, select
-from sqlalchemy.ext.asyncio import AsyncSession
+from fastapi import Depends, HTTPException
 
-from dependencies import get_db
-
-from posts.models import Post, PostView
-from posts.repository import PostRepository, PostSortEnum, get_post_repo
+from src.posts.models import Post, PostView, PostSortEnum
+from src.posts.repository import PostRepository, get_post_repo
 
 
 class PostService:
     def __init__(
             self, 
             repo:PostRepository=Depends(get_post_repo), 
-            db:AsyncSession=Depends(get_db), 
         ):
         self.repo = repo
-        self.db = db
 
     async def create_post(self, user_id:int, title:str, content:str):
-        logger.logger.debug('create_post service')
         new_post_view = PostView()
         new_post = Post(user_id=user_id, title=title, content=content, post_view=new_post_view)
-        created_post = await self.repo.save(new_post)
-        return created_post
+
+        return await self.repo.save(new_post)
 
     async def update_post(self, user_id:int, post_id: int, title:str, content:str):
-        post = await self.get_post(post_id)
-
+        post = await self.repo.find_post_by_id(post_id)
+        if post is None:
+            raise HTTPException(status_code=400, detail='존재하지 않는 포스트입니다.')
+        
         if post.user_id != user_id:
             raise HTTPException(status_code=401, detail='작성자 본인만 수정가능 합니다.')
 
@@ -38,39 +32,48 @@ class PostService:
         if content:
             post.update_content(content)
 
-        try:
-            result = await self.repo.update(user_id, post)
-            return result
-        except Exception as e:
-            print(e)
-            raise HTTPException(status_code=500, detail='회원정보 수정이 실패했습니다. 기입한 내용을 확인해보세요')
+        return await self.repo.update(user_id, post)
+        
 
     async def delete_post(self, user_id:int, post_id: int):
-        post = await self.get_post(post_id)
+        post = await self.repo.find_post_by_id(post_id)
+        if post is None:
+            raise HTTPException(status_code=400, detail='존재하지 않는 포스트입니다.')
+        
         if post.user_id != user_id:
             raise HTTPException(status_code=401, detail='작성자 본인만 삭제가능 합니다.')
         
         await self.repo.delete(user_id, post_id)
-        return post.id
+        return post
 
-    async def get_posts(self, page: int = 1, items_per_page: int = 20, sort_option: Optional[PostSortEnum] = PostSortEnum.CREATED_AT):
-        posts = await self.repo.find_all(page, items_per_page, sort_option)
-        # dict화를 중심으로 분리할까?
-        def post_form_changer(post: Post):
-            post_dict = post.__dict__
-            post_dict['username'] = post.user.username
-            post_dict['view_count'] = post.post_view.view_count
-            return post_dict
-        return list(map(lambda post: post_form_changer(post), posts))
+    async def get_posts(
+            self, 
+            page: int = 1, 
+            items_per_page: int = 20, 
+            sort_option: Optional[PostSortEnum] = PostSortEnum.CREATED_AT
+        ):
+        finded_posts = await self.repo.find_all(page, items_per_page, sort_option)
 
+        posts = self._post_form_changer(finded_posts)
+        return posts
+
+    def _post_form_changer(self, finded_posts:List[Post]):
+        for post in finded_posts:
+            username = self._username_selector(post)
+            post.add_username(username)
+            post.add_view_count(post.post_view.view_count)
+        return finded_posts
+    
+    def _username_selector(self, post:Post):
+        deleted_username = '탈퇴한 유저'
+        return post.user.username if post.user.deleted_at is None else deleted_username
+    
     async def get_post(self, post_id: int) -> Post:
         post = await self.repo.get_post_detail(post_id)
         if post is None:
             raise HTTPException(status_code=400, detail='존재하지 않는 포스트입니다.')
         
-        # 이게 맞나?
-        post.username = post.user.username
-
+        post.add_username(post.user.username)
         return post
 
 def get_post_service(service: PostService = Depends(PostService)):
